requirements.txt
# GENERATED FROM requirements.in.  DO NOT EDIT DIRECTLY.
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
beautifulsoup4==4.9.3
boto3==1.17.14
botocore==1.20.14
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.7
django-admin-autocomplete-filter==0.6.1
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-extensions==3.1.1
django-import-export==2.5.0
django-localflavor==3.0.1
django-object-actions==3.0.1
django-rest-swagger==2.2.0
django-storages==1.11.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
ecdsa==0.14.1
et-xmlfile==1.0.1
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
html2text==2020.1.16
httplib2==0.19.0
idna==2.10
ipython==7.20.0
ipython-genutils==0.2.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupPy==1.14
MarkupSafe==1.1.1
nested-lookup==0.2.22
num2words==0.5.10
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.6
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
protobuf==3.15.2
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
python-jose==3.2.0
python-stdnum==1.16
pytz==2021.1
PyYAML==5.4.1
raven==6.10.0
requests==2.25.1
rsa==4.7.2
s3transfer==0.3.4
simplejson==3.17.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tablib==3.0.0
traitlets==5.0.5
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
yapf==0.30.0
